requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
Flask==0.12.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pkg-resources==0.0.0
python-craigslist==1.0.5
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.4
urllib3==1.22
Werkzeug==0.14.1
